fix(transform): Use chunk fileName to determine transformed module name for relative `declare module` blocks
The original code used to determine the transformed module name in `declare module` blocks with a relative reference used the `file` option passed to rollup with a default fallback. This does not work when that option is not provided, e.g. when `dir` is used instead or when multiple chunks are generated.

The fix determines the target module name based on the `fileName` property of the current chunk, instead.